Reorganize the code structure to allow unit tests to run
Moved the main logic to its own package.
The test can import the main logic using absolute imports.

Also refactored some imports.
